Take the seek origin back at pointer width

+++ code/vqalib/dstream.cpp
@@ -43,6 +43,7 @@
 ****************************************************************************/
 
 #include	"vqaplayp.h"
+#include	<stdint.h>
 #include	<stdio.h>
 #include	<fcntl.h>
 #include	<io.h>
@@ -98,17 +99,17 @@ long __cdecl Disk_VQA_Stream_Handler(VQAHandle *vqa, long action, void *buffer, 
 		 * VQAERR_SEEK.
 		 */
 		case VQACMD_SEEK:
-			error = (lseek(fh, nbytes, (long)buffer) == -1);
+			error = (lseek(fh, nbytes, (int)(intptr_t)buffer) == -1);
 			break;
 
 		case VQACMD_SEEKPEEK:
 			if (nbytes > 0) {
-				error = lseek(fh, nbytes - 1, (int)buffer) == -1;
+				error = lseek(fh, nbytes - 1, (int)(intptr_t)buffer) == -1;
 				if (error == 0) {
 					error = read(fh, &temp, 1) != 1;
 				}
 			} else {
-				error = lseek(fh, nbytes, (int)buffer) == -1;
+				error = lseek(fh, nbytes, (int)(intptr_t)buffer) == -1;
 				if (error == 0) {
 					error = read(fh, &temp, 1) != 1;
 				}
@@ -196,7 +197,7 @@ long __cdecl Memory_VQA_Stream_Handler(VQAHandle *vqa, long action, void *buffer
 		 */
 		case VQACMD_SEEK:
 		case VQACMD_SEEKPEEK:
-			switch ((long)buffer) {
+			switch ((int)(intptr_t)buffer) {
 
 				case 1:
 					cache->Offset += nbytes;
